chore(table): cleanup demo by moving to examples
