--- src/demo-app/table/custom-table/custom-table.scss
@@ -1,9 +1,0 @@
-.mat-table, wrapper-table {
-  width: 100%;
-}
-
-.mat-card {
-  margin: 8px 0;
-  height: 300px;
-  overflow: auto;
-}
