--- src/demo-app/table/data-input-table/data-input-table.scss
@@ -1,12 +1,0 @@
-.demo-actions {
-  margin-top: 24px;
-}
-
-.demo-track-by-select {
-  display: flex;
-  align-items: center;
-
-  .demo-track-by-label {
-    margin-right: 8px;
-  }
-}
