--- src/demo-app/table/table-demo.html
@@ -1,7 +1,0 @@
-<material-example-list [ids]="examples"></material-example-list>
-
-<custom-table-demo></custom-table-demo>
-
-<data-input-table-demo></data-input-table-demo>
-
-<when-rows-demo></when-rows-demo>
