+++ src/material-examples/table-simple-column/table-simple-column-example.css
@@ -1,0 +1,3 @@
+table {
+  width: 100%;
+}
