+++ src/material-examples/table-simple-column/table-simple-column-example.html
@@ -1,0 +1,15 @@
+<table mat-table [dataSource]="dataSource" class="mat-elevation-z8"
+       matSort #sort="matSort">
+  <!-- Basic column: name is used for header label AND data property -->
+  <simple-column name="name" sortable></simple-column>
+  <simple-column name="position"></simple-column>
+
+  <!-- Name doesn't match the data property (or transform needed); define a custom data accessor -->
+  <simple-column name="weight" [dataAccessor]="getWeight"></simple-column>
+
+  <!-- Name doesn't match desired header text; define a custom label -->
+  <simple-column name="symbol" label="SYMBOL!"></simple-column>
+
+  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+  <tr mat-row *matRowDef="let data; columns: displayedColumns;"></tr>
+</table>
