+++ src/material-examples/table-simple-column/table-simple-column-example.ts
@@ -1,16 +1,52 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import {coerceBooleanProperty} from '@angular/cdk/coercion';
 import {Component, Input, OnDestroy, OnInit, Optional, ViewChild} from '@angular/core';
-import {MatSortHeader} from '@angular/material/sort';
-import {MatColumnDef, MatTable} from '@angular/material/table';
+import {coerceBooleanProperty} from '@angular/cdk/coercion';
+import {
+    MatColumnDef,
+    MatSort,
+    MatSortHeader,
+    MatTable,
+    MatTableDataSource
+} from '@angular/material';
 
+export interface PeriodicElement {
+  name: string;
+  position: number;
+  weight: number;
+  symbol: string;
+}
+
+const ELEMENT_DATA: PeriodicElement[] = [
+  {position: 1, name: 'Hydrogen', weight: 1.0079, symbol: 'H'},
+  {position: 2, name: 'Helium', weight: 4.0026, symbol: 'He'},
+  {position: 3, name: 'Lithium', weight: 6.941, symbol: 'Li'},
+  {position: 4, name: 'Beryllium', weight: 9.0122, symbol: 'Be'},
+  {position: 5, name: 'Boron', weight: 10.811, symbol: 'B'},
+  {position: 6, name: 'Carbon', weight: 12.0107, symbol: 'C'},
+  {position: 7, name: 'Nitrogen', weight: 14.0067, symbol: 'N'},
+  {position: 8, name: 'Oxygen', weight: 15.9994, symbol: 'O'},
+  {position: 9, name: 'Fluorine', weight: 18.9984, symbol: 'F'},
+  {position: 10, name: 'Neon', weight: 20.1797, symbol: 'Ne'},
+];
+
+/**
+ * @title Table with a custom column component for easy column definition reuse.
+ */
+@Component({
+  selector: 'table-simple-column-example',
+  styleUrls: ['table-simple-column-example.css'],
+  templateUrl: 'table-simple-column-example.html',
+})
+export class TableSimpleColumnExample implements OnInit {
+  displayedColumns: string[] = ['position', 'name', 'weight', 'symbol'];
+  dataSource = new MatTableDataSource<PeriodicElement>(ELEMENT_DATA);
+  getWeight = (data: PeriodicElement) => '~' + data.weight;
+
+  @ViewChild('sort') sort: MatSort;
+
+  ngOnInit() {
+    this.dataSource.sort = this.sort;
+  }
+}
 
 /**
  * Column that shows simply shows text content for the header and row
