+++ src/material-examples/table-wrapped/table-wrapped-example.css
@@ -1,0 +1,3 @@
+table {
+  width: 100%;
+}
